Tab -> Spaces

+++ src/lib/DirectoryRecord.h
@@ -4,7 +4,7 @@
 #include "INode.h"
 
 struct DirectoryRecord {
-	uint32_t length;
-	INode::ID inode_ID;
-	char name[];
+  uint16_t length;
+  INode::ID inode_ID;
+  char name[];
 };
